fix(ci): python-coverage-shards - split oversized status test part
